+++ ICD.Connect.Calendaring/CalendarPoints/ICalendarPoint.cs
@@ -1,17 +1,10 @@
-﻿using ICD.Common.Properties;
-using ICD.Connect.Calendaring.Controls;
+﻿using ICD.Connect.Calendaring.Controls;
 using ICD.Connect.Devices.Points;
 
 namespace ICD.Connect.Calendaring.CalendarPoints
 {
-	public interface ICalendarPoint : IPoint
+	public interface ICalendarPoint : IPoint<ICalendarControl>
 	{
-		/// <summary>
-		/// Gets the control for this point.
-		/// </summary>
-		[CanBeNull]
-		new ICalendarControl Control { get; }
-
 		/// <summary>
 		/// Gets/Sets the features for this point.
 		/// </summary>